+++ src/pages/AsistenteInvestigacion/EstadisticasInvestigaciones/CantidadTrabajosXAutor.js
@@ -4,13 +4,21 @@
 import React, {useState,useEffect, Component } from 'react';
 import { Grid } from '@mui/material';
 import { Bar } from 'react-chartjs-2';
-import BarraDocumentosXAutor from './BarraDocumentosXAutor';
+import BarraDocumentosXAutor from '../../../components/PageComponents/BarCharts';
 import InvestigacionService from '../../../services/investigacionService';
 
 const autoresIndicadores = async () => {
     let indicadoresAutores = await InvestigacionService.getIndicadoresAutores();
     console.log(indicadoresAutores)
     return indicadoresAutores;
+}
+
+const estandarizarAutoresInd = arr => {
+    let arrEstandarizado=[];
+    arr.forEach(element => {
+        arrEstandarizado.push([element[0].nombres+element[0].apellidos,element[1]]);
+    });
+    return arrEstandarizado;
 }
 
 export default function CantidadTrabajosXAutor(){
@@ -22,14 +30,15 @@ export default function CantidadTrabajosXAutor(){
         .then(newAutorIndicador => {
             setAutoresInd(newAutorIndicador);
         });
+        setAutoresInd(estandarizarAutoresInd(autoresInd));
     }, [])
+
+    /* {BarraDocumentosXAutor.BarChartAutores()} */
 
     return(
 
         <div>
-
-            {console.log(autoresInd)}
-
+           {console.log(autoresInd)}
         </div>
 
     )
